refactor: rename file-casing in ui pkg and web
